Improve ergonomics and add instance tags support
Now the user can optionally customize the tags put on the created
instance, so customers in environments that require a specific set of
tags can still utilize the script.

Also improves the use of user-friendly names for created resources, and
outputs a URL for each analysis result for easier access to detailed
results.