Add `EnableDql` parameter
This parameter will probably be useful for AWS Backup Restore testing scenario where we don't need anything fancy, especially a DQL. This should reduce the CFN deployment time a bit.

Also moved the logic for uploading the CFN and replacing the placeholders to `upload.sh` so that it is usable for local development.