Fix trying to get the `SubnetId` of a terminated instance
Terminated instances have no Vpc/Subnet info in their DescribeInstances response. The code was trying to figure out if there were any cleanups scheduled for the vpc+subnet pair to avoid running cleanup in them, however we can't get that info from DescribeInstances. The workaround is to uze the AZ info that is still available in the instance to consider all vpc+subnet pairs within the same AZ to potentially contain the terminated instance and thus delay cleanup in them.

Note that this may lead to a problem where there are several VPCs and there is a lot of activity in one VPC, but no activity in the other. In this case cleanup may be delayed indefinitely in the VPC without activity. A proper solution to this problem would involve storing the VPC/subnet state somewhere, for example instance tags, which would require more time to implement.